Made Transition Keys Exclusive
Error with the dictionary key for transitions not being able to be
exclusive such that two of the same key could exist. Obviously this is
not the proper use of a dictionary so it was changed.

Started work on designing the lines coming off of states to represent
transitions. More math is required