State Transition Design
Completed the math required to have the transition lines not overlap on
a to and from transition to the same states.

Added some GUI images, not quite happy with the result but it's what
I've got.

TODO: Dictionaries do not serialize and therefore do not work with
scriptable objects. A workaround is to serialize and deserialize the
dictionary yourself. Use dictionary when manipulating data and store
them into multiple lists afterward...doesn't sound fun, but may be the
solution. Another work around is needed for usability.