End of Day Commit
Converted most of the code to use all scriptable objects for states and
transitions. Not going smoothly currently as I'm not sure if states
should store transitions or if there should be a master list with all
transitions or even have states store to and from transitions
separately. It gets messy when things get deleted and lists need to sync
up with the newly deleted data.