Select Transitions
Transitions can now be selected in the window by clicking on the little
arrow (which is not an object but rather a position I stored in a rect)
and will turn blue while selected. I would allow for clicking along the
line as well, but that's far too many calculations for me.