Transition Limits and Entry Point
Added some limits to the type of transitions that a state allows and the
amount it allows. Added an entry point variable for the FSM to hold so
that it can keep track of what is currently set as the initial state of
the FSM.

Cleaned some destruction code to fix potential future errors.

Running into a null reference error after compile, but it doesn't affect
being able to use the window. Not sure what's up with that.