Dynamic State Machine
Renamed files and added a folder structure with namespaces to represent
the actual function of the class. It is not a finite state machine as
the amount and name of states could potentially change even at runtime
if I so desired to allow it. It is a dynamic state machine instead. Now
its name reflects it. Some other name changes were made for clarity sake
such as the name for 'ScriptableFSM' was changed to 'DSMObject' to match
other scriptableobjects such as 'DSMState'