IAmUsingInterfaces
Converted all scriptableobjects to implement interfaces as well. While
it still isn't completely modular, it at least makes the editor window
for it modular. The editor window can handle a IDSMObject, IDSMStates,
and IDSMTransitions that implement the corresponding interfaces.

Added the ability to transition from state to state by state by the
state's display name. Since state names change in the editor just like
in mecanim, it creates a lot of overhead to rename states for the user.
Not sure of a better solution than what unity already implements.