AutoProperty Property Drawers
Tested to make sure unity functions worked properly.

Added property drawers for all of the common types of primitive types
and vectors. Creating an auto property and hooking it up to a callback
is simple. Viewing it in the Unity inspector is also simple.

Need to figure out a way to serialize and display the Unity Event
nativly if possible. This would expand the usefulness of this class 10
fold in Unity.